T2: fix csv import

+++ src/main/java/com/solbegsoft/city/library/model/Role.java
@@ -9,7 +9,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
-import java.math.BigDecimal;
 
 @Entity
 @Data
@@ -19,7 +18,7 @@ public class Role {
     @Id
     @SequenceGenerator(name = "role_id_seq", sequenceName = "role_id_seq", allocationSize = 1)
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "role_id_seq")
-    private BigDecimal id;
+    private Long id;
 
     @Column
     private String name;
